% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddNameStat}
\alias{AddNameStat}
\title{Function will update the name with the statistic of your choice}
\usage{
AddNameStat(df, category, count_col, stat = c("sd", "mean", "count"),
  dp = 0)
}
\arguments{
\item{df}{a dataframe}

\item{category}{the category being named}

\item{count_col}{the column being aggregated against}

\item{stat}{the statistics. "sd" is standard deviation, "mean" or "count".}

\item{dp}{decimal places of returned value}
}
\description{
Function will update the name with the statistic of your choice
}
