% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ChiSquared}
\alias{ChiSquared}
\title{Calculate Chi squared statistics}
\usage{
ChiSquared(Model)
}
\arguments{
\item{Model}{A glm object}
}
\description{
Internal function used to calculates the Chi Squared statistics for a model
}
