% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CrossValidation}
\alias{CrossValidation}
\title{Function for internally validating a glm model.}
\usage{
CrossValidation(dataframe, iterations, foldSize, fullstats = FALSE)
}
\arguments{
\item{dataframe}{the dataframe which contains the parameters to be modelled. The first column must contain the outcome variable}

\item{iterations}{the number of iterations for the function to run}

\item{foldSize}{the size of the test dataset expressed as a value between 0 and 1}

\item{fullstats}{TRUE/FALSE. Should the full stats be exported?}
}
\description{
Function for internally validating a glm model.
}
