% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DurbinWatsonCheck}
\alias{DurbinWatsonCheck}
\title{Calculates Durbin-Watson statistic for glm model}
\usage{
DurbinWatsonCheck(Model)
}
\arguments{
\item{Model}{A glm object}
}
\description{
Calculate the Durbin-Watson statistic to detect the presence of
 autocorrelation in the residuals from a regression analysis.
}
