% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grViz_pdf}
\alias{grViz_pdf}
\title{# Save a file from DiagrammeR as a PDF figure which can be included in a knitr report.
Saves PDF output in the same directory as the ".gv" file specified.}
\usage{
grViz_pdf(filepath, saveSuffix = "")
}
\arguments{
\item{filepath}{The input location of the .gz file}

\item{saveSuffix}{An optional term to be added to the end of the PDF.}
}
\description{
# Save a file from DiagrammeR as a PDF figure which can be included in a knitr report.
Saves PDF output in the same directory as the ".gv" file specified.
}
