% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loessfit}
\alias{loessfit}
\title{Fit a LOESS curve to a dataset}
\usage{
loessfit(x, DegreeValue, SpanValue)
}
\arguments{
\item{x}{the parameter name for the model}

\item{DegreeValue}{the degree of polynomials to be used}

\item{SpanValue}{the parameter ?? which controls the degree of smoothing}
}
\description{
Fit a LOESS curve to a dataset
}
