% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticModel}
\alias{LogisticModel}
\title{Function to build logistic model from a list of input parameters}
\usage{
LogisticModel(variables, df)
}
\arguments{
\item{variables}{a list of predictor variables}

\item{df}{model dataframe}
}
\description{
Function to build logistic model from a list of input parameters
}
