% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticOddsPlot}
\alias{LogisticOddsPlot}
\title{Function builds the}
\usage{
LogisticOddsPlot(Model, Sort = FALSE, Title)
}
\arguments{
\item{Model}{a glmmodel}

\item{Sort}{indicates whether the parameters should be sorted}

\item{Title}{a character string of the title of the of ggplot}
}
\description{
Function builds the
}
