% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogResultsTable}
\alias{LogResultsTable}
\title{Takes a list of logistic regression models and creates a summary table}
\usage{
LogResultsTable(...)
}
\arguments{
\item{...}{a list of glm objects}
}
\description{
Takes a list of logistic regression models and creates a summary table
}
