% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matchCategory}
\alias{matchCategory}
\title{Matches variable names with their full descriptive name.
Used when plotting or showing results in a table}
\usage{
matchCategory(inputNames, path = "VariableDisplayNames.csv")
}
\arguments{
\item{inputNames}{a list of names which are to looked up}

\item{path}{the filepath of the lookup file}
}
\description{
This function is designed to work with the PhD project
}
