% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ModelAccuracy}
\alias{ModelAccuracy}
\title{Assesses the accuracy of a logistic regression model}
\usage{
ModelAccuracy(dataframe, outcomeVariable, predictorVariables,
  iterations = 300, foldSize = 0.05, fullstats = FALSE)
}
\arguments{
\item{dataframe}{a data frame}

\item{outcomeVariable}{a list of the outcome variables}

\item{predictorVariables}{a string of the column name for the outcome variable}

\item{iterations}{number of iterations to run for the model accuracy. Defaults 300}

\item{foldSize}{size of data proportion used to assess model accuracy. Defaults 0.05}

\item{fullstats}{TRUE/FALSE. Should the full results be returned? Default FALSE}
}
\description{
Assesses the accuracy of a logistic regression model
}
