% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my.summary}
\alias{my.summary}
\title{Produces a summary of a column within a dataframe}
\usage{
my.summary(x, rounded = 1, ...)
}
\arguments{
\item{x}{a vector of values}

\item{rounded}{the number of decimal places to return}

\item{...}{additional arguments parsed by summary statistic calculations}
}
\value{
a list of summary statistics for the variable
}
\description{
Produces a summary of a column within a dataframe
}
