% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OddsTable}
\alias{OddsTable}
\title{Build Odds Table}
\usage{
OddsTable(Model, round = 3)
}
\arguments{
\item{Model}{A glm object}

\item{round}{the number of decimal places the figure should be printed to}
}
\description{
Creates an odds table for parameters within the regression model
}
