% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{A colour scale fill for the publication}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{option arguments passed to `discrete_scale`}
}
\description{
A colour scale fill for the publication
}
