% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_FullOddsPlot}
\alias{Segmented_FullOddsPlot}
\title{Splits a dataset and forms a segmented Odds plot}
\usage{
Segmented_FullOddsPlot(df, split_category, predictors, outcome, Title,
  linebreak)
}
\arguments{
\item{df}{a dataframe to be split}

\item{split_category}{the column to be used to split the dataset by}

\item{predictors}{a list of variables to be used within the regression model}

\item{outcome}{the outcome variable}

\item{Title}{a string to be used within the ggplot title}

\item{linebreak}{the spacing to be used between gridlines in the plot}
}
\description{
This function splits data based on a categorical variable, builds separate
  glm models for each dataset and compares the results using a Odds Plot
}
