% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_OddsTable}
\alias{Segmented_OddsTable}
\title{Builds Segmented Regression Odds Tables}
\usage{
Segmented_OddsTable(LogisticModelList)
}
\arguments{
\item{LogisticModelList}{a list of logistic regression models}
}
\value{
a single odds table which contains the statistics for the segmented model
}
\description{
Split dataset into segments and build logistic models into a list
}
