% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SummariseDataframe}
\alias{SummariseDataframe}
\title{Summarise a Dataframe}
\usage{
SummariseDataframe(dataframe)
}
\arguments{
\item{dataframe}{a datatable}

\item{...}{additional arguments parsed to `my.sumarry`}
}
\description{
Produces summary statistics for rows in a dataframe
}
